# play/chess-tourney-server/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "(title)",
    "description": "API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/bots/create": {
      "post": {
        "operationId": "Bots_create",
        "summary": "Create a new bot user - Admin only",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBot"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            }
          },
          "500": {
            "description": "ISE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                }
              }
            }
          }
        },
        "tags": [
          "Bot Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBot"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/bots/delete/{botId}": {
      "delete": {
        "operationId": "Bots_delete",
        "summary": "Delete a bot user (changes the user type back to Regular) - Admin only",
        "parameters": [
          {
            "name": "botId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            }
          },
          "500": {
            "description": "ISE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                }
              }
            }
          }
        },
        "tags": [
          "Bot Management"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/bots/update/{botId}": {
      "patch": {
        "operationId": "Bots_update",
        "summary": "Update an existing bot's configuration - Admin",
        "parameters": [
          {
            "name": "botId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBot"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            }
          },
          "500": {
            "description": "ISE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                }
              }
            }
          }
        },
        "tags": [
          "Bot Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBot"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "BadRequestError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "Bad Request"
      },
...
      "ConflictError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "Conflict"
      },
      "CreateBot": {
        "type": "object",
        "required": [
          "username",
          "playouts_per_move",
          "exploration_factor",
          "virtual_loss",
          "contempt",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3
          },
          "model_key": {
            "type": "string"
          },
          "playouts_per_move": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "exploration_factor": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 20
          },
          "virtual_loss": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5
          },
          "contempt": {
            "type": "number",
            "format": "float",
            "minimum": -3,
            "maximum": 3
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "format": "password"
          }
        },
        "description": ""
      },
...
      "ForbiddenError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "Forbidden"
      },
...
      "InternalError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "ISE"
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "Not Found"
      },
...
      "ReadBot": {
        "type": "object",
        "required": [
          "user_id",
          "bot_id",
          "username",
          "playouts_per_move",
          "exploration_factor",
          "virtual_loss",
          "contempt"
        ],
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "bot_id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "minLength": 3
          },
          "model_key": {
            "type": "string"
          },
          "playouts_per_move": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "exploration_factor": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 20
          },
          "virtual_loss": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5
          },
          "contempt": {
            "type": "number",
            "format": "float",
            "minimum": -3,
            "maximum": 3
          }
        },
        "description": ""
      },
...
      "UnauthorizedError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "description": "Unauthorized access"
      },
      "UpdateBot": {
        "type": "object",
        "required": [
          "username",
          "playouts_per_move",
          "exploration_factor",
          "virtual_loss",
          "contempt"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3
          },
          "model_key": {
            "type": "string"
          },
          "playouts_per_move": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "exploration_factor": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 20
          },
          "virtual_loss": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 5
          },
          "contempt": {
            "type": "number",
            "format": "float",
            "minimum": -3,
            "maximum": 3
          }
        },
        "description": ""
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "Bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ]
}